added nft link to yt view page; refactored out the nft link element
added yt content script
