+++ src/extension/Content/Selectors/youtube.ts
@@ -1,0 +1,1 @@
+export const VideoCountSelector = 'ytd-video-view-count-renderer'
