+++ src/manifest.json
@@ -2,7 +2,9 @@
   "manifest_version": 3,
   "name": "Chrome Extension for <place_holder>",
   "options_page": "options.html",
-  "background": { "service_worker": "background.bundle.js" },
+  "background": {
+    "service_worker": "background.bundle.js"
+  },
   "action": {
     "default_popup": "popup.html",
     "default_icon": "icon-34.png"
@@ -12,15 +14,35 @@
   },
   "content_scripts": [
     {
-      "matches": ["http://*.tiktok.com/*", "https://*.tiktok.com/*"],
-      "js": ["contentScript.bundle.js"],
-      "css": ["content.styles.css"]
+      "matches": [
+        "http://*.tiktok.com/*",
+        "https://*.tiktok.com/*"
+      ],
+      "js": [
+        "contentScript.bundle.js"
+      ],
+      "css": [
+        "content.styles.css"
+      ]
+    },
+    {
+      "matches": [
+        "https://*.youtube.com/*",
+        "http://*.youtube.com/*"
+      ],
+      "js": [
+        "ytContentScript.bundle.js"
+      ]
     }
   ],
   "devtools_page": "devtools.html",
   "web_accessible_resources": [
     {
-      "resources": ["content.styles.css", "icon-128.png", "icon-34.png"],
+      "resources": [
+        "content.styles.css",
+        "icon-128.png",
+        "icon-34.png"
+      ],
       "matches": []
     }
   ]
